requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0
celery>=5.3.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
